Renamed all consoles.
Since the idea of consoles is a conceptual one, renamed all current consoles,
and their associated types, to stations and station_types. Basically, the idea
is that of a human sitting at a station that does a particular function, not the
idea of a bash console, or its associated idea called a terminal. The idea is in
the functionality. To use Star Trek as an example, Data and Geordi sat at the
Nav Station. Spock stood at the Science Station. Underneath that idea of a Station
sat the idea of a console, namely the physical device they tapped on to do their jobs,
but the collection of the parts and functions that console did was under the connotation
of a Station. That is what the API is putting forth here; a collection of toolsets
collectively called the Station. The Chair where Captain Picard sat, with all of it's own
toolsets and capabilities, would be the API's idea of a MasterStation. Hopefully, that explains
the reasoning behind the rename; to better reflect what the cognitive connotations are meant
when we discuss a Station vs. a Console.